requirements.txt
iso639-lang==2.1.0
wordfreq==3.0.3
ipadic==1.0.0
jieba==0.42.1
mecab-ko-dic==1.0.0
mecab-python3==1.0.6
PyStemmer==2.0.1
